requirements.txt
Flask~=3.1
scikit-learn~=1.6
flask-cors~=5.0.1
sqlalchemy~=2.0.39
pyjwt~=2.10.1
bcrypt~=4.3.0
pydantic~=2.10.6
python-dotenv~=1.1.0
